Allow .bundle markers to signal that a bundle should provide resources for a given namespace
As to the utility of this feature, consider the example of the
clojure.core.incubator namespace.

As this is loaded from clojure/core/incubator.clj, it is considered --
for OSGi purposes -- to live in the clojure.core package; thus, attempts
to load it are routed to the bundle providing clojure.core, confounding
attempts to package and version these independently-developed tools
separately.

This can be resolved by having a jar containing
clojure/core/incubator/.bundle, and advertising itself as providing the
clojure.core.incubator package (with its own version). Should such a
.bundle marker be found by the clojure.osgi loading mechanism, only the
jar in which the .bundle marker was found will be used for related class
or .clj files.